+++ ignite/templates/ibc/placeholders.go
@@ -2,28 +2,11 @@ package ibc
 
 //nolint:godot
 const (
-	Placeholder  = "// this line is used by starport scaffolding # 1"
-	Placeholder2 = "// this line is used by starport scaffolding # 2"
-	Placeholder3 = "// this line is used by starport scaffolding # 3"
+	Placeholder = "// this line is used by starport scaffolding # 1"
 
 	// Placeholders IBC packets
 	PlaceholderIBCPacketEvent         = "// this line is used by starport scaffolding # ibc/packet/event"
 	PlaceholderIBCPacketModuleRecv    = "// this line is used by starport scaffolding # ibc/packet/module/recv"
 	PlaceholderIBCPacketModuleAck     = "// this line is used by starport scaffolding # ibc/packet/module/ack"
 	PlaceholderIBCPacketModuleTimeout = "// this line is used by starport scaffolding # ibc/packet/module/timeout"
-
-	// Placeholders for messages
-	PlaceholderProtoTxRPC     = "// this line is used by starport scaffolding # proto/tx/rpc"
-	PlaceholderProtoTxMessage = "// this line is used by starport scaffolding # proto/tx/message"
-
-	// Placeholders for Oracle
-	PlaceholderProtoTxImport          = "// this line is used by starport scaffolding # proto/tx/import"
-	PlaceholderOraclePacketModuleRecv = "// this line is used by starport scaffolding # oracle/packet/module/recv"
-	PlaceholderOraclePacketModuleAck  = "// this line is used by starport scaffolding # oracle/packet/module/ack"
-	PlaceholderOracleModuleRecv       = "// this line is used by starport scaffolding # oracle/module/recv"
-	PlaceholderOracleModuleAck        = "// this line is used by starport scaffolding # oracle/module/ack"
-
-	// Placeholders AutoCLI
-	PlaceholderAutoCLIQuery = "// this line is used by ignite scaffolding # autocli/query"
-	PlaceholderAutoCLITx    = "// this line is used by ignite scaffolding # autocli/tx"
 )