+++ ignite/templates/message/placeholders.go
@@ -2,6 +2,5 @@ package message
 
 const (
 	Placeholder  = "// this line is used by starport scaffolding # 1"
-	Placeholder2 = "// this line is used by starport scaffolding # 2"
 	Placeholder3 = "// this line is used by starport scaffolding # 3"
 )