+++ ignite/templates/module/create/base.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/ignite/cli/v28/ignite/pkg/gomodulepath"
 	"github.com/ignite/cli/v28/ignite/pkg/placeholder"
+	"github.com/ignite/cli/v28/ignite/pkg/xast"
 	"github.com/ignite/cli/v28/ignite/pkg/xgenny"
 	"github.com/ignite/cli/v28/ignite/pkg/xstrings"
 	"github.com/ignite/cli/v28/ignite/templates/field/plushhelpers"
@@ -81,17 +82,29 @@ func appConfigModify(replacer placeholder.Replacer, opts *CreateOptions) genny.R
 		}
 
 		// Import
-		template := `%[2]vmodulev1 "%[3]v/api/%[4]v/%[2]v/module"
-_ "%[3]v/x/%[2]v/module" // import for side-effects
-%[2]vmoduletypes "%[3]v/x/%[2]v/types"
-%[1]v`
-		replacement := fmt.Sprintf(template, module.PlaceholderSgAppModuleImport, opts.ModuleName, opts.ModulePath, opts.AppName)
-		content := replacer.Replace(fConfig.String(), module.PlaceholderSgAppModuleImport, replacement)
+		content, err := xast.AppendImports(
+			fConfig.String(),
+			xast.WithLastNamedImport(
+				fmt.Sprintf("%[1]vmodulev1", opts.ModuleName),
+				fmt.Sprintf("%[1]v/api/%[2]v/%[3]v/module", opts.ModulePath, opts.AppName, opts.ModuleName),
+			),
+			xast.WithLastNamedImport(
+				"_",
+				fmt.Sprintf("%[1]v/x/%[2]v/module", opts.ModulePath, opts.ModuleName),
+			),
+			xast.WithLastNamedImport(
+				fmt.Sprintf("%[1]vmoduletypes", opts.ModuleName),
+				fmt.Sprintf("%[1]v/x/%[2]v/types", opts.ModulePath, opts.ModuleName),
+			),
+		)
+		if err != nil {
+			return err
+		}
 
 		// Init genesis
-		template = `%[2]vmoduletypes.ModuleName,
+		template := `%[2]vmoduletypes.ModuleName,
 %[1]v`
-		replacement = fmt.Sprintf(template, module.PlaceholderSgAppInitGenesis, opts.ModuleName)
+		replacement := fmt.Sprintf(template, module.PlaceholderSgAppInitGenesis, opts.ModuleName)
 		content = replacer.Replace(content, module.PlaceholderSgAppInitGenesis, replacement)
 		replacement = fmt.Sprintf(template, module.PlaceholderSgAppBeginBlockers, opts.ModuleName)
 		content = replacer.Replace(content, module.PlaceholderSgAppBeginBlockers, replacement)
@@ -138,15 +151,21 @@ func appModify(replacer placeholder.Replacer, opts *CreateOptions) genny.RunFn {
 		}
 
 		// Import
-		template := `%[2]vmodulekeeper "%[3]v/x/%[2]v/keeper"
-%[1]v`
-		replacement := fmt.Sprintf(template, module.PlaceholderSgAppModuleImport, opts.ModuleName, opts.ModulePath)
-		content := replacer.Replace(f.String(), module.PlaceholderSgAppModuleImport, replacement)
+		content, err := xast.AppendImports(
+			f.String(),
+			xast.WithLastNamedImport(
+				fmt.Sprintf("%[1]vmodulekeeper", opts.ModuleName),
+				fmt.Sprintf("%[1]v/x/%[2]v/keeper", opts.ModulePath, opts.ModuleName),
+			),
+		)
+		if err != nil {
+			return err
+		}
 
 		// Keeper declaration
-		template = `%[2]vKeeper %[3]vmodulekeeper.Keeper
+		template := `%[2]vKeeper %[3]vmodulekeeper.Keeper
 %[1]v`
-		replacement = fmt.Sprintf(
+		replacement := fmt.Sprintf(
 			template,
 			module.PlaceholderSgAppKeeperDeclaration,
 			xstrings.Title(opts.ModuleName),
@@ -155,14 +174,18 @@ func appModify(replacer placeholder.Replacer, opts *CreateOptions) genny.RunFn {
 		content = replacer.Replace(content, module.PlaceholderSgAppKeeperDeclaration, replacement)
 
 		// Keeper definition
-		template = `&app.%[2]vKeeper,
-		%[1]v`
-		replacement = fmt.Sprintf(
-			template,
-			module.PlaceholderSgAppKeeperDefinition,
-			xstrings.Title(opts.ModuleName),
+		content, err = xast.ModifyFunction(
+			content,
+			"New",
+			xast.AppendInsideFuncCall(
+				"Inject",
+				fmt.Sprintf("\n&app.%[1]vKeeper", xstrings.Title(opts.ModuleName)),
+				-1,
+			),
 		)
-		content = replacer.Replace(content, module.PlaceholderSgAppKeeperDefinition, replacement)
+		if err != nil {
+			return err
+		}
 
 		newFile := genny.NewFileS(appPath, content)
 		return r.File(newFile)