+++ ignite/templates/module/placeholders.go
@@ -3,13 +3,10 @@ package module
 //nolint:godot
 const (
 	Placeholder  = "// this line is used by starport scaffolding # 1"
-	Placeholder2 = "// this line is used by starport scaffolding # 2"
 	Placeholder3 = "// this line is used by starport scaffolding # 3"
 
 	// Placeholders in app.go
-	PlaceholderSgAppModuleImport      = "// this line is used by starport scaffolding # stargate/app/moduleImport"
 	PlaceholderSgAppKeeperDeclaration = "// this line is used by starport scaffolding # stargate/app/keeperDeclaration"
-	PlaceholderSgAppKeeperDefinition  = "// this line is used by starport scaffolding # stargate/app/keeperDefinition"
 	PlaceholderSgAppInitGenesis       = "// this line is used by starport scaffolding # stargate/app/initGenesis"
 	PlaceholderSgAppBeginBlockers     = "// this line is used by starport scaffolding # stargate/app/beginBlockers"
 	PlaceholderSgAppEndBlockers       = "// this line is used by starport scaffolding # stargate/app/endBlockers"
@@ -17,7 +14,6 @@ const (
 	PlaceholderSgAppModuleConfig      = "// this line is used by starport scaffolding # stargate/app/moduleConfig"
 
 	// Placeholders IBC
-	PlaceholderIBCImport    = "// this line is used by starport scaffolding # ibc/app/import"
 	PlaceholderIBCKeysName  = "// this line is used by starport scaffolding # ibc/keys/name"
 	PlaceholderIBCKeysPort  = "// this line is used by starport scaffolding # ibc/keys/port"
 	PlaceholderIBCNewModule = "// this line is used by starport scaffolding # ibc/app/module"